Fix 既 vs 旣 distinction as a component
U+65E2 既 has Compatibility Ideograph U+FA42 既,
which is extremely dumb, since U+65E3 旣 should have it instead.
This commit prevents this dumbness from being inherited
when 既 and 旣 are used as components.
In particular, note that it is U+69EA 槪 (rather than U+6982 概)
which contains the (U+FA42 既)-like component in Shinjitai.